People app created
