+++ acm_web_site/apps/business/people/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from . import views
 
 
-handler404 = 'apps.business.web_site.views.page_not_found'
+handler404 = 'apps.business.people.views.page_not_found'
 
 urlpatterns = [
 	path('members', views.members),
